Redesign reviewMain() and add tolerance customization
I took apart and rewrote the entirety of the primary function for
this script, as I wasn't satisfied with the way the original aVars
object was instantiated nor handled throughout. In its place is a new
`mrmr` object which houses basically everything not already stored in
the settings, including functions.

In addition, there is now support for the following customizations:

- Refuse 'Close' Answers: instead of accepting 'within tolerance'
 typos, the typo is refused and prompts to try again.

- Accuracy Tolerance: offers three modes (relaxed, moderate, strict)
 which range from WaniKani's default leniency to almost no leniency.
 On 'strict,' only very long answers will allow any typos.