Add alert messages, rewrite checkAnswer() and answerChecker()
Removed evalAnswer() function, splitting its actions between
checkAnswer() and the modified answerChecker(). Rewrote entirety
of both functions for easier maintenance and clear order of ops.
Changed the formatting of html and css text so it doesn't use tons
of apostrophes and plus signs.

Added the following functions:
- findMatch
- setExceptionStatus
- setPassingStatus
- setAlert
- loadQuizItems
- findItem

The first three new functions are used in answerChecker() to find
the exact/closest match from answer arrays, set true/false exception
status and type of exception, and final pass/fail of answer.

If an exception is set, setAlert() sets the appropriate text to be
used in the additional mutationObserver (specifically looks for
WK's popup message) to replace the default text.

loadQuizItems() uses wkof.ItemData.get_items function. The only
purpose for this is for allowing a reading exception on single-
character vocab question, similar to onyomi/kunyomi exceptions for
kanji questions. This is needed because I don't otherwise know how
to access the item's component's data directly from jStorage (which
means only seeing the kana, and not the emphasis required).

The findItem() function is used only when the current question item
is a single-character vocab. It searches the object array returned
by loadQuizItems() for the matching object by its 'id' property. It's
run twice, once to get the vocab's object (and thus its component's
id), and second to get the component object which is then used to
determine which of the kanji's reading is required for the vocab.

STILL TO DO:

The new mutationObserver replaces the text as it should, but the
fadeOut (or animate, which I also tried) functions aren't changing
the opacity like they should. For now I've just set it to be removed
after 3 seconds.

Finish testing all possible combinations of settings.